+++ crates/gpui_platform/src/lib.rs
@@ -1,3 +1,5 @@
 pub use current_platform::*;
 #[cfg(any(test, feature = "test-support"))]
 mod tests;
+#[cfg(any(test, feature = "test-support"))]
+use gpui_macros::test;
