+++ crates/gpui_platform/src/tests/elements/mod.rs
@@ -1,0 +1,3 @@
+mod label;
+mod list;
+mod text;
